Allow to run warning passes on model with erroneous value instances
Currently the pass DeriveValueInstanceTypes throws error if type of instance
cannot be calculated, but this will change in next commit. That means that
accessing `dataType` of the ValueInstanceSpec will throw an error. If such
error is thrown we just cannot analyze that instance and can just ignore it

(review this commit in whitespace changes ignored mode)