meta/ks-version: allow only versions 0.6 and higher
As explained in
https://github.com/kaitai-io/ksy_schema/commit/44987cf11806eae1d58c1e253c28d015c20a200d,
it does not make sense to specify any lower `ks-version` than 0.6, since
KS 0.6 is the first version that accepts the `meta/ks-version` key at
all (older versions like KS 0.5 would complain as `/meta/ks-version:
unknown key found, ...`).

Most likely no one would want to deliberately specify such versions.
However, version `0.1` is particularly likely to have been created by
mistake, because that is how YAML parses the naive specification of
`ks-version: 0.10` due to `0.10` matching the float number syntax and
`0.10 == 0.1` for float numbers. Therefore, we generate an extra
parenthesis with a hint to use `ks-version: '0.10'` if we detect that we
have received `0.1`.