GH Actions: add `fail-fast: false`
Otherwise there is no point in having a matrix strategy: we currently
have 53 failing tests, so the resulting status of each CI job is
"failed". With `fail-fast: true` (default setting), only one job is
completed (i.e. one version of Java; usually Java 8, but it's not
deterministic) and the others are always cancelled.