Remove broken and unused CircleCI pipeline
It's just a waste of CPU resources, because we don't need it for
anything. We run tests in GitHub Actions (and it seems that `sbt test`
in this CircleCI pipeline is broken anyway for some reason, it just
prints `[success] Total time: 0 s` - maybe it's running in the wrong
directory). Building and publishing a universal .zip snapshot build of
KSC is AFAIK currently not done anywhere, but it seems that this doesn't
work either:

```
Uploading /home/circleci/repo/jvm/target/universal/kaitai-struct-compiler-*.zip to jvm/target/universal/kaitai-struct-compiler-*.zip
  No artifact files found at /home/circleci/repo/jvm/target/universal/kaitai-struct-compiler-*.zip
Total size uploaded: 0 B
```