GH Actions: remove matrix strategy, test on Java 21 only
It seems highly unlikely to me that there would be differences in
behavior between Java versions. I don't think we use any APIs that
differ between Java versions. So what are we actually testing here? Not
so much our code, but perhaps whether Scala and the libraries we use
work across Java versions. That doesn't make much sense to me. Let's
just trust that their maintainers are doing their job and that they work
on all LTS versions of Java.

Furthermore, if the purpose was really to verify that the compiler works
across JVM versions (including both our code and the libraries we use),
then end-to-end tests would be much more appropriate than these unit
tests, which only test a small part of a few isolated places in the
compiler's codebase.