Stop checking `_parent` and `_root` of external types
Since https://github.com/kaitai-io/kaitai_struct_compiler/pull/283, it's
incorrect to assume that `_parent` should be `this` and `_root` should
be `this._root` for external types. Instead, when crossing the .ksy spec
boundary to an external type `obj`, `obj._root == obj` and
`obj._parent == null` should hold, but we will not check that for now.