Stop avoiding the local variable `i` in `WriteBackHandler`
Demonstrates https://github.com/kaitai-io/kaitai_struct/issues/1246 in
both Python and Java. Therefore, this commit breaks ~10 KST tests (so
you can actually reproduce the problem, if you like), but they will be
fixed in the next commit.

I think it's better to use the `i` variable than to refer to the last
element as `self._raw_blocks[-1]` or
`this._raw_blocks.get(this._raw_blocks.size() - 1)`, because it leads to
simpler and more consistent code.