Remove `_check{Inst}()` methods, check instances in `_check()`
If I remember correctly, the reason parse instances were not checked in
the main `_check()` method was the concern that not all instances are
intended for writing. Examples of instances that are typically not meant
to be written are lookahead instances or inactive `union` members (that
are not guarded with `if` in the .ksy spec). Since such instances are
not supposed to be written, there is no point in checking their
consistency - they should be treated as if they did not exist at all.

Now that we have the `set{Inst}_ToWrite` / `{inst}__to_write` mechanism
for the serialization user to tell Kaitai Struct which instances to
ignore when writing, I think it makes perfect sense to also use these
flags to decide whether to check each instance for consistency in
`_check()`.

This change makes serialization much easier to use (correctly/as
intended; of course, nothing currently forces you to call any `_check()`
methods, but then you're missing out on all the consistency checks),
because users no longer have to explicitly call all `_check{Inst}()`
methods for each instance separately (when there is actually no reason
why these checks could not be included in `_check()`).