Java: rename the private `_write*` flags to `_shouldWrite*`
This is the same name already used in Python, where it would otherwise
clash with the `_write{Inst}()` method. I think it's good to keep names
consistent across languages and `_shouldWrite*` is arguably a better
name anyway (naming the `_write{Inst}()` method and the corresponding
boolean flag the same may have been confusing).