ZigCompiler: remove helper method `fromFile`
I'm not sure if we'll ever have this method in Zig, because reading from
a file apparently implies having to manage the `std.fs.File` object, the
`std.fs.File.Reader` and its buffer. I guess we could allocate these
dynamically using the passed allocator, and add a `close()` method to
KaitaiStream that would close the file, but it seems quite weird. Either
way, it's not an essential functionality, so let's not worry about it
for now.