Implement initial support for Zig (instances not supported)
Many of the KST tests (100+) are already passing in Zig with this
commit. However, most notably, instances do not work yet, mainly due to
"error: declarations are not allowed between container fields". Fixing
this error will require touching the language-agnostic code, because Zig
requires all fields in the struct to be declared together (see
https://github.com/ziglang/zig/issues/5077), as opposed to interleaving
the instance caching field and getter as we do now.