Zig: don't declare `_raw_*` fields, use local variables instead
Currently, the KS-generated code almost always exposes `_raw_*` fields
publicly, but I don't see much value in that. In Zig, this complicates
things when compiling
https://github.com/kaitai-io/kaitai_struct_tests/blob/f681ced791cd882beefecf7d3576743b4ea451c2/formats/instance_user_array.ksy#L19-L25
(because the `_raw_*` field was declared as an optional type, but used
as non-optional in the instance getter), so I'm switching to local
variables in Zig specifically. Eventually, we may decide to do the same
for other target languages.