C++: fix `assignType` for byte array "else" case in switch
The `assignType` parameter should represent the type of the variable
we're assigning to. `RawIdentifier(id)` is always a byte array field, so
using `Some(assignType)` as in the other branches does not reflect
reality.

This change has no effect on code generation because currently the only
target with `switchBytesOnlyAsRaw` is C++, which does not care about
`assignType` in this case. But it's still a mistake.