Allow SwitchType to be passed as `assignType`
This has no effect on the generated code yet (unless I've overlooked
something). However, it allows us to implement
https://github.com/kaitai-io/kaitai_struct/issues/1265 in Zig, because
as you can see in
https://github.com/kaitai-io/kaitai_struct/issues/1265#issuecomment-3633481378,
we have to wrap the value in a tagged union explicitly, like this:

```diff
                 34 => {
                     const _raw_body = try self._io.readBytes(self._allocator(), self.size);
                     const _io__raw_body = try self._allocator().create(_imp_kaitai_struct.KaitaiStream);
                     _io__raw_body.* = _imp_kaitai_struct.KaitaiStream.fromBytes(_raw_body);
-                    self.body = try ChunkDir.create(self._arena, _io__raw_body, self, self._root);
+                    self.body = Body_switch{ .chunk_dir = try ChunkDir.create(self._arena, _io__raw_body, self, self._root) };
                 },
```

To do this, we obviously need to know whether the struct field we are
assigning to uses a tagged union or not.